requirements.txt
pytesseract>=0.3.10
pillow>=10.0.0
python-xlib>=0.33
pynput>=1.7.6
requests==2.24.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
mss>=9.0.1
numpy>=1.24.0
opencv-python>=4.8.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
urllib3>=1.26.17
websockets>=12.0